newCommit

+++ src/main/java/ru/sitnikov/personal/personal/model/UserModel.java
@@ -16,7 +16,7 @@ public class UserModel {
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "user_seq_generator")
     @SequenceGenerator(name = "user_seq_generator", sequenceName = "user_seq", allocationSize = 1)
-    private Integer id;
+    private Long id;
     @Column(name = "name", nullable = false)
     private String name;
 
@@ -35,12 +35,12 @@ public class UserModel {
     @Column(name = "login")
     private String login;
 
-    public UserModel(Integer id, String name) {
+    public UserModel(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public UserModel(Integer id, String name, String lastName) {
+    public UserModel(Long id, String name, String lastName) {
         this.id = id;
         this.name = name;
         this.lastName = lastName;
